test5/testnet-training-node-quickstart/requirements.txt
torch>=1.13.1
transformers>=4.37.2
peft>=0.10.0
loguru
trl>=0.9.3
bitsandbytes
pyyaml
accelerate